Input Manager and Skills
Added improvments to the input manager. It now checks for both computer
and controller inputs.
Can also choose whether to check exclusively to make sure one is pressed
and not the other for onKeyDown or onKeyUp events.

Started the skill system
Need to implement the skill effects to work

Object pool can now check if a behaviour script has been added to the
object as well as adding it if it does not have it